+++ src/App.js
@@ -1,78 +1,15 @@
-import React, {useState, useRef, useMemo, useEffect} from "react";
-import {PostList} from "./components/PostList";
+import React from "react";
 import "./styles/App.css";
-import {PostForm} from "./components/PostForm";
-import {PostFilter} from "./components/PostFilter";
-import {MyModal} from "./components/UI/MyModal/MyModal";
-import {MyButton} from "./components/UI/button/MyButton";
-import {usePosts} from "./hooks/NewPosts";
-import PostService from "./API/PostService";
-import Loader from "./components/UI/Loader/Loader";
-import {useFetching} from "./hooks/useFetching";
-import {getPageCount, getPagesArray} from "./utils/pages";
-import Pagintaion from "./components/UI/pagination/Pagintaion";
+import {BrowserRouter} from "react-router-dom";
+import Navbar from "./components/UI/Navbar/Navbar";
+import AppRouter from "./components/AppRouter";
 
 export function App() {
-    const [posts, setPosts] = useState([]);
-    const [filter, setFilter] = useState({sort: '', query: ''})
-    const [modal, setModal] = useState(false);
-    const [totalPages, setTotalPages] = useState(0);
-    const [limit, setLimit] = useState(10);
-    const [page, setPage] = useState(1);
-    const sortedAndSearchedPosts = usePosts(posts, filter.sort, filter.query);
-    const [fetchPosts, isPostsLoading, postError] = useFetching(async () => {
-        const response = await PostService.getAll(limit, page);
-        setPosts(response.data);
-        const totalCount = (response.headers['x-total-count'])
-        setTotalPages(getPageCount(totalCount, limit));
-    })
-
-    useEffect(()=>{
-        fetchPosts()
-    }, [page])
-
-    const createPost = (newPost) => {
-        setPosts([...posts, newPost])
-        setModal(false)
-    }
-
-    const removePost = (post) => {
-        setPosts(posts.filter(p => p.id !== post.id))
-    }
-
-    const changePage = (page) => {
-        setPage(page);
-    }
-
     return (
-        <div className="App">
-            <MyButton style={{marginTop: 30}} onClick={() => setModal(true)}>
-                Создать пост
-            </MyButton>
-            <MyModal visible={modal} setVisible={setModal}>
-                <PostForm create={createPost}/>
-            </MyModal>
-
-            <hr style={{margin: '15px 0'}}/>
-            <PostFilter
-                filter={filter}
-                setFilter={setFilter}
-            />
-            {postError &&
-                <h1>произошла ошибка ${postError}</h1>
-            }
-            {isPostsLoading
-                ? <div style={{display: 'flex', justifyContent: 'center', marginTop: 50}}>
-                    <Loader/>
-                  </div>
-                :<PostList remove={removePost} posts={sortedAndSearchedPosts} title={"Список постов"}/>
-            }
-            <Pagintaion
-                page={page}
-                changePage={changePage}
-                totalPages={totalPages}
-            ></Pagintaion>
-        </div>
+        <BrowserRouter>
+            <Navbar />
+            <AppRouter />
+        </BrowserRouter>
     )
 }
