+++ src/components/PostForm.jsx
@@ -1,4 +1,4 @@
-import React, {useState, useRef} from "react";
+import React, {useState} from "react";
 import { MyButton } from "../components/UI/button/MyButton";
 import { MyInput } from "../components/UI/input/MyInput"
