+++ src/pages/About.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const About = () => {
+    return (
+        <h1>
+            Это приложение создано
+        </h1>
+    );
+};
+
+export default About;
