+++ src/pages/Error.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Error = () => {
+    return (
+        <div>
+            Page not found
+        </div>
+    );
+};
+
+export default Error;
